Context menu devtools follow the devtools docking preference state
