+++ desktop-app/src/main/webview-context-menu/register.ts
@@ -1,5 +1,5 @@
 import { BrowserWindow, ipcMain, Menu } from 'electron';
-import { CONTEXT_MENUS } from './handler';
+import { CONTEXT_MENUS } from './common';
 
 export const initWebviewContextMenu = () => {
   ipcMain.on('show-context-menu', (event, ...args) => {
